Bug Fixes
Had an outdated and incorrect killall in the uninstaller

Added 2 sleeps back in, one for the post install start and one for the auto start on login.  This should help with the window failing to start properly or instantly crashing.  Also modified the indentation for the writes to the autorun script