Bug FIxes
Fixed the bug where when ran. an error could occur if X Server was not ready in the specified time frame.  Mitigated the need for sleep timers.  --kiosk is no longer an option when installed, has been pushed to the actual run command which is now prox-kiox as a bash command.  Removed Openbox as a dependancy and added X11 Utils which is used now for checking if X Server is running, more reliable, stabler.  Prettied up the installer.